Move setup logic to new file. Implement function that finds legal setup positions.
